# EES-3004 - optimised matching on FilterItemIds in table results stored procedures