--- a/src/GovUk.Education.ExploreEducationStatistics.Admin/appsettings.json
+++ b/src/GovUk.Education.ExploreEducationStatistics.Admin/appsettings.json
@@ -29,7 +29,7 @@
     "TokenValidationParameters.RoleClaimType": "role"
   },
   "TableBuilder": {
-    "MaxTableCellsAllowed": 250000
+    "MaxTableCellsAllowed": 25000
   },
   "Locations": {
     "TableResultLocationHierarchiesEnabled": false,