--- a/src/GovUk.Education.ExploreEducationStatistics.Data.Model/Migrations/20220104144102_EES3004_UpdateSelectObservationsRoutinesOptimiseFilterItemIds.cs
+++ b/src/GovUk.Education.ExploreEducationStatistics.Data.Model/Migrations/20220104144102_EES3004_UpdateSelectObservationsRoutinesOptimiseFilterItemIds.cs
@@ -1,0 +1,29 @@
+﻿using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
+using Microsoft.EntityFrameworkCore.Migrations;
+using static GovUk.Education.ExploreEducationStatistics.Data.Model.Migrations.MigrationConstants;
+
+namespace GovUk.Education.ExploreEducationStatistics.Data.Model.Migrations
+{
+    public partial class EES3004_UpdateSelectObservationsRoutinesOptimiseFilterItemIds : Migration
+    {
+        private const string MigrationId = "20220104144102";
+
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.SqlFromFile(MigrationsPath, 
+                $"{MigrationId}_Routine_SelectObservations.sql");
+            
+            migrationBuilder.SqlFromFile(MigrationsPath, 
+                $"{MigrationId}_Routine_SelectObservationsByLocationCodes.sql");
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.SqlFromFile(MigrationsPath, 
+                $"{EES2778_QueryObservationsByLocationId.MigrationId}_Routine_SelectObservations.sql");
+            
+            migrationBuilder.SqlFromFile(MigrationsPath, 
+                $"{EES2778_QueryObservationsByLocationId.MigrationId}_Routine_SelectObservationsByLocationCodes.sql");
+        }
+    }
+}