--- a/src/GovUk.Education.ExploreEducationStatistics.Data.Model/Migrations/20220104144102_Routine_SelectObservations.sql
+++ b/src/GovUk.Education.ExploreEducationStatistics.Data.Model/Migrations/20220104144102_Routine_SelectObservations.sql
@@ -36,7 +36,7 @@ DECLARE
                 
     IF (@filterItemsExist = 1)
         SET @sql += N'JOIN ObservationFilterItem ofi ON o.Id = ofi.ObservationId ' +
-                     'JOIN #FilterItemId filterItemId ON ofi.FilterItemId = filterItemId.id '
+                     'AND ofi.FilterItemId IN (SELECT id FROM #FilterItemId) '
 
     SET @sql += N'WHERE o.SubjectId = @subjectId ' +
                  'AND o.LocationId IN (' + @locationIdsList + ') '