This PR:
- optimises the 2 existing "SelectObservations" stored procedures with a couple of changes.

This is in response to problems we're having on Prod with Subjects like QL01, where certain Filters can have thousands of potential FilterItem choices available.  In the case of QL01, around 10,000 choices exist for the Qualification filter for example when all Time Periods are chosen.

One change is to replace IN clauses that handle large numbers of values from table variables with IN clauses on a temporary table instead.  It was not possible to convert the list of FilterItemIds to a csv string of FilterItemIds using STRING_AGG and then use that string of ids in an IN clause, as is used widely in the `SelectObservationsByLocationCodes` stored proc for other lists of ids and codes, because the query engine cannot handle a SQL-stringed IN clause with several thousand possible values in it.  5,000 was OK, but 10,000 caused it to run out of resources.

Another change is the optimisation to the HAVING clause, whereby it now checks that the number of matching FilterItems per Observation row is equal to the number of Filters that that are present in the set of FilterItemIds without having to do several JOINs to other tables in the process.

Note that we might also want to make a similar change to remove the IN clause for Location ids in the new `SelectObservation` query, replacing it with IN clauses on a temporary table of LocationIds, but this'll come with testing when the Location Id support is all in.

The results of these changes as tested manually on pre-prod's public statistics db with a QL01 table tool query brought a 240+ seconds query down to 6-20 seconds.

Capturing before-and-after datablock results on pre-prod and then comparing them with the command `diff -I"^time for response:.*" -I "Not Found" -r results_preprod_before/ results_preprod_after/` shows that the results were the same, and the overall run time reduced from 2100 seconds to 1033 seconds to query 120+ datablocks when public-statistics was unscaled at 1 vCore.

Note also that:
* The original variant of this stored procedure had no index on the temp table and was using a JOIN on the temporary table.  Its run time was 1900 seconds.
* The original variant of this stored procedure plus the index on the temp table and the JOIN on the temporary table's run time was 1200 seconds.

Therefore all variants of the original were faster, but the best performing was with the index on the temp table, and the IN clause against the temp table rather than the JOIN. 

![image](https://user-images.githubusercontent.com/12444316/148409274-2489b988-30fa-4c85-9549-6b07b5e13469.png)